refactor: slash command params
