+++ packages/app/app/components/Message/Model/Message/Input/Header/SlashCommandParameters.vue
@@ -1,109 +1,17 @@
 <script setup lang="ts">
-import type { SlashCommandParameters } from "@/models/message/slashCommands/SlashCommandParameters";
-import type { SubmitEventPromise } from "vuetify";
-import type { VForm } from "vuetify/components";
-
-import { SlashCommandType } from "@/models/message/slashCommands/SlashCommandType";
-import { sanitizeHtml } from "@/services/sanitizeHtml/sanitizeHtml";
-import { formRules } from "@/services/vuetify/formRules";
-import { useDataStore } from "@/store/message/data";
 import { useSlashCommandStore } from "@/store/message/input/slashCommand";
-import { useRoomStore } from "@/store/message/room";
-import { MessageType } from "@esposter/db-schema";
-import { marked } from "marked";
 
-const roomStore = useRoomStore();
-const { currentRoomId } = storeToRefs(roomStore);
 const slashCommandStore = useSlashCommandStore();
-const { parameterValues, pendingSlashCommand } = storeToRefs(slashCommandStore);
+const { pendingSlashCommand } = storeToRefs(slashCommandStore);
 const { clearPendingSlashCommand } = slashCommandStore;
-const dataStore = useDataStore();
-const { createMessage } = dataStore;
-const formRef = useTemplateRef<InstanceType<typeof VForm>>("formRef");
-const isFormValid = ref(true);
-const { execute, isLoading } = useInFlight();
-const submit = (event: SubmitEventPromise) =>
-  execute(async () => {
-    const { valid } = await event;
-    if (!valid || !pendingSlashCommand.value || !currentRoomId.value) return;
-    const command = { parameterValues: parameterValues.value, type: pendingSlashCommand.value.type } as {
-      [K in SlashCommandType]: { parameterValues: SlashCommandParameters<K>; type: K };
-    }[SlashCommandType];
-    const roomId = currentRoomId.value;
-
-    switch (command.type) {
-      case SlashCommandType.Me:
-        await createMessage({
-          message: marked.parse(`*${sanitizeHtml(command.parameterValues.message)}*`, { async: false }),
-          roomId,
-          type: MessageType.Message,
-        });
-        break;
-      case SlashCommandType.Shrug: {
-        const prefix = command.parameterValues.text?.trim() ?? "";
-        await createMessage({
-          message: marked.parse(`${prefix}¯\\\\\\_(ツ)\\_/¯`, { async: false }),
-          roomId,
-          type: MessageType.Message,
-        });
-        break;
-      }
-    }
-
-    clearPendingSlashCommand();
-  });
-
-useEventListener("keydown", (event: KeyboardEvent) => {
-  if (event.key === "Escape") clearPendingSlashCommand();
-});
 </script>
 
 <template>
-  <div v-if="pendingSlashCommand" w-full>
-    <MessageModelMessageInputHeader @close="clearPendingSlashCommand()">
-      <v-icon icon="mdi-slash-forward" size="small" />
-      <span>
-        <span font-bold>{{ pendingSlashCommand.title }}</span>
-        <span opacity-60 ml-1>{{ pendingSlashCommand.description }}</span>
-      </span>
-      <StyledEditFormDialogErrorIcon :edit-form="formRef ?? undefined" :is-edit-form-valid="isFormValid" />
-    </MessageModelMessageInputHeader>
-    <v-form ref="formRef" v-model="isFormValid" @submit.prevent="submit">
-      <StyledCard>
-        <div flex items-start gap-2 px-4 py-3>
-          <template v-for="{ description, isRequired, name } of pendingSlashCommand.parameters" :key="name">
-            <v-text-field
-              v-model="parameterValues[name]"
-              :placeholder="isRequired ? name : `${name} (optional)`"
-              :label="description"
-              :rules="isRequired ? [formRules.required] : []"
-              density="compact"
-              variant="plain"
-              hide-details
-              autofocus
-              flex-1
-            />
-          </template>
-          <v-tooltip text="Send (Enter)">
-            <template #activator="{ props: tooltipProps }">
-              <v-btn
-                type="submit"
-                icon="mdi-send"
-                size="small"
-                density="compact"
-                variant="text"
-                mt-2
-                :disabled="isLoading"
-                :loading="isLoading"
-                :="tooltipProps"
-              />
-            </template>
-          </v-tooltip>
-        </div>
-      </StyledCard>
-    </v-form>
-    <div flex justify-between px-1 pt-1>
-      <MessageModelMessageInputFooter />
-    </div>
-  </div>
+  <MessageModelMessageInputHeader v-if="pendingSlashCommand" @close="clearPendingSlashCommand()">
+    <v-icon icon="mdi-slash-forward" size="small" />
+    <span>
+      <span font-bold>{{ pendingSlashCommand.title }}</span>
+      <span opacity-60 ml-1>{{ pendingSlashCommand.description }}</span>
+    </span>
+  </MessageModelMessageInputHeader>
 </template>
